requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
resend
pydantic[email]
python-dotenv
